Move emojis to its own class
